separate rss flight
