+++ src/main/kotlin/com/emilburzo/service/rss/rss.kt
@@ -1,6 +1,6 @@
 package com.emilburzo.service.rss
 
-import com.emilburzo.service.Flight
+import com.emilburzo.service.RssFlight
 import com.rometools.rome.io.SyndFeedInput
 import com.rometools.rome.io.XmlReader
 import org.slf4j.LoggerFactory
@@ -13,10 +13,10 @@ class Rss {
 
     private val logger = LoggerFactory.getLogger(javaClass)
 
-    fun getFlights(content: String): List<Flight> {
+    fun getFlights(content: String): List<RssFlight> {
         val result = SyndFeedInput().build(XmlReader(content.byteInputStream()))
         logger.info("found ${result.entries.size} flights")
-        return result.entries.map { mapFlight(it.title, it.link) }
+        return result.entries.map { mapRssFlight(it.title, it.link) }
     }
 
 }
